solved some grafical bugs
git-svn-id: https://xp-dev.com/svn/biappi-HandcraftSW@100 0f4557bd-cc6f-0410-bf7d-5d4a72267dac

+++ Autovelox/trunk/Autovelox2Pro/Classes/TutorAlertDetailsView.m
@@ -89,7 +89,7 @@
 	}
 	else 
 	{
-		dis.text=[NSString stringWithFormat:@"%1.1f",(float)(distanzaFineTutor/1000)];
+		dis.text=[NSString stringWithFormat:@"%1.1f",(float)((float)distanzaFineTutor/1000)];
 		unitDis.text=[NSString stringWithFormat:@"Km"];
 	}
 	//rilascia
